Laid out main level
main level is now laid out.  The Lobby will load to the main level
instead of the test level. The furniture are marked with a tag Furniture
and have colliders on them so that players cant run through them.  items
also now spawn based on the size of the map (public field in
pickupSpawner)  Also fixed camera sprite to render (name was previously
misspelled and therefor didn't render